Only request the entities for a surface once.
Remember the set of entities for a given surface instead of calling
surface.find_entities() #pole_entities + 1 times. When fetching the set
of entities for a surface that are powered, exclude entities named
'tf-pole' from the initial request, to further optimize
script.on_configuration_changed. These changes allow an early-ish
AngelBobs save game to load in under a minute instead of trying to load
for 20+ minutes and triggering the out-of-memory killer on a Linux
computer with 32GB of RAM (Factorio would hit 15GB used after 5-10 min
of trying to load the save game).